+++ lib/src/widgets/reactive_form_field.dart
@@ -15,17 +15,19 @@ class ReactiveFormField<T> extends StatefulWidget {
   /// Function that returns the widget representing this form field. It is
   /// passed the form field state as input, containing the current value and
   /// validation state of this field.
-  final ReactiveFormFieldBuilder<T> builder;
+  final ReactiveFormFieldBuilder<T> _builder;
   final String formControlName;
   final Map<String, String> validationMessages;
 
   const ReactiveFormField({
     Key key,
     @required this.formControlName,
-    @required this.builder,
+    @required ReactiveFormFieldBuilder<T> builder,
     Map<String, String> validationMessages,
   })  : assert(formControlName != null),
+        assert(builder != null),
         validationMessages = validationMessages ?? const {},
+        _builder = builder,
         super(key: key);
 
   @override
@@ -34,12 +36,23 @@ class ReactiveFormField<T> extends StatefulWidget {
 
 class ReactiveFormFieldState<T> extends State<ReactiveFormField<T>> {
   FormControl control;
+  bool _touched;
 
   /// The current value of the [FormControl].
   T get value => this.control.value;
 
+  bool get touched => _touched;
+
+  set touched(bool value) {
+    if (this._touched != value) {
+      setState(() {
+        this._touched = value;
+      });
+    }
+  }
+
   String get errorText {
-    if (this.control.invalid && this.control.touched) {
+    if (this.control.invalid && this.touched) {
       return widget.validationMessages
               .containsKey(this.control.errors.keys.first)
           ? widget.validationMessages[this.control.errors.keys.first]
@@ -52,7 +65,7 @@ class ReactiveFormFieldState<T> extends State<ReactiveFormField<T>> {
   @override
   void initState() {
     this.control = _getFormControl();
-    this.subscribeFormControl();
+    this.subscribeControl();
 
     super.initState();
   }
@@ -61,9 +74,9 @@ class ReactiveFormFieldState<T> extends State<ReactiveFormField<T>> {
   void didChangeDependencies() {
     final newControl = _getFormControl();
     if (this.control != newControl) {
-      this.unsubscribeFormControl();
+      this.unsubscribeControl();
       this.control = newControl;
-      subscribeFormControl();
+      subscribeControl();
     }
 
     super.didChangeDependencies();
@@ -71,20 +84,24 @@ class ReactiveFormFieldState<T> extends State<ReactiveFormField<T>> {
 
   @override
   void dispose() {
-    this.unsubscribeFormControl();
+    this.unsubscribeControl();
     super.dispose();
   }
 
   @protected
-  void subscribeFormControl() {
-    this.control.onStatusChanged.addListener(_onFormControlStatusChanged);
-    this.control.onValueChanged.addListener(_onFormControlValueChanged);
+  void subscribeControl() {
+    this.control.onStatusChanged.addListener(_onControlStatusChanged);
+    this.control.onValueChanged.addListener(_onControlValueChanged);
+    this.control.onTouched.addListener(_onControlTouched);
+
+    this._touched = this.control.touched;
   }
 
   @protected
-  void unsubscribeFormControl() {
-    this.control.onStatusChanged.removeListener(_onFormControlStatusChanged);
-    this.control.onValueChanged.removeListener(_onFormControlValueChanged);
+  void unsubscribeControl() {
+    this.control.onStatusChanged.removeListener(_onControlStatusChanged);
+    this.control.onValueChanged.removeListener(_onControlValueChanged);
+    this.control.onTouched.removeListener(_onControlTouched);
   }
 
   FormControl _getFormControl() {
@@ -94,37 +111,40 @@ class ReactiveFormFieldState<T> extends State<ReactiveFormField<T>> {
       throw FormControlParentNotFoundException(widget);
     }
 
-    return form.formControl(widget.formControlName);
+    return form.control(widget.formControlName);
   }
 
-  void _onFormControlValueChanged() {
+  void _onControlValueChanged() {
     this.updateValueFromControl();
   }
 
   @protected
   void updateValueFromControl() {
-    this.touch();
+    touch();
   }
 
-  void _onFormControlStatusChanged() {
+  @protected
+  void touch() {
+    this.control.touch();
+  }
+
+  void _onControlStatusChanged() {
     setState(() {});
+  }
+
+  void _onControlTouched() {
+    this.touched = this.control.touched;
   }
 
   void didChange(T value) {
     this.control.value = value;
-    if (this.control.touched) {
+    if (this.touched) {
       setState(() {});
     }
   }
 
-  void touch() {
-    setState(() {
-      this.control.touched = true;
-    });
-  }
-
   @override
   Widget build(BuildContext context) {
-    return widget.builder(this);
+    return widget._builder(this);
   }
 }